+++ CHANGELOG.md
@@ -20,6 +20,7 @@ Control panel:
 
 Other:
 
+* Set systemd journald log retention to 10 days (from no limit) to reduce disk usage.
 * Fixed log processing for submission lines that have a sasl_sender or other extra information.
 
 Version 55 (October 18, 2021)